GEODE-3893: Replace mocks utility class for all examples with per-example mocks
