+++ listener/src/test/java/org/apache/geode/examples/listener/ExampleTest.java
@@ -19,8 +19,6 @@ import static org.junit.Assert.assertEquals;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.apache.geode.cache.Region;
-import org.geode.examples.util.Mocks;
 import org.junit.Test;
 
 public class ExampleTest {
